#68: Pre-release code revise:  - updated RP client to 4.0.13  - moved recorders to separate package  - removed tree optimised preset  - updated javadoc
