+++ pom.xml
@@ -60,7 +60,7 @@
         <dependency>
             <groupId>com.epam.reportportal</groupId>
             <artifactId>client-java</artifactId>
-            <version>4.0.11</version>
+            <version>4.0.13</version>
         </dependency>
         <dependency>
             <groupId>com.epam.reportportal</groupId>
