+++ src/main/java/com/github/invictum/reportportal/ReportPortalListener.java
@@ -1,6 +1,7 @@
 package com.github.invictum.reportportal;
 
 import com.github.invictum.reportportal.injector.IntegrationInjector;
+import com.github.invictum.reportportal.recorder.TestRecorder;
 import com.google.inject.Inject;
 import net.thucydides.core.model.DataTable;
 import net.thucydides.core.model.Story;
@@ -52,90 +53,90 @@ public class ReportPortalListener implements StepListener {
     }
 
     public void testRetried() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void stepStarted(ExecutedStepDescription description) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void skippedStepStarted(ExecutedStepDescription description) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void stepFailed(StepFailure failure) {
-        collectDriverLogs();
+        harvestDriverLogs();
     }
 
     public void lastStepFailed(StepFailure failure) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void stepIgnored() {
-        collectDriverLogs();
+        harvestDriverLogs();
     }
 
     public void stepPending() {
-        collectDriverLogs();
+        harvestDriverLogs();
     }
 
     public void stepPending(String message) {
-        collectDriverLogs();
+        harvestDriverLogs();
     }
 
     public void stepFinished() {
-        collectDriverLogs();
+        harvestDriverLogs();
     }
 
     public void testFailed(TestOutcome testOutcome, Throwable cause) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void testIgnored() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void testSkipped() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void testPending() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void testIsManual() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void notifyScreenChange() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void useExamplesFrom(DataTable table) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void addNewExamplesFrom(DataTable table) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void exampleStarted(Map<String, String> data) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void exampleFinished() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void assumptionViolated(String message) {
-        /* Not used by listener */
+        // Not used by listener
     }
 
     public void testRunFinished() {
-        /* Not used by listener */
+        // Not used by listener
     }
 
-    private void collectDriverLogs() {
+    private void harvestDriverLogs() {
         if (ThucydidesWebDriverSupport.isDriverInstantiated()) {
             Logs logs = ThucydidesWebDriverSupport.getDriver().manage().logs();
             logStorage.collect(logs);
