+++ src/test/java/com/github/invictum/reportportal/LogsPresetTest.java
@@ -31,10 +31,4 @@ public class LogsPresetTest {
         LogsPreset preset = LogsPreset.CUSTOM.register(Essentials.startStep());
         Assert.assertEquals(1, preset.logUnits().length);
     }
-
-    @Test
-    public void treeOptimizedProfile() {
-        int actual = LogsPreset.TREE_OPTIMIZED.logUnits().length;
-        Assert.assertEquals(2, actual);
-    }
 }
